set token context on login
