+++ web/src/components/auth/tokenContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const AuthTokenContext = createContext("");
