+++ web/src/components/auth/utils.js
@@ -2,13 +2,13 @@ import axios from "axios";
 import { apiUrlAuthLogin } from "../../api.util";
 
 /**
- * 
- * @param {*} email 
- * @param {*} password 
- * @returns 
+ *
+ * @param {*} email
+ * @param {*} password
+ * @returns
  */
 export const login = async (email, password) => {
-  let token = '';
+  let token = "";
   let code = 100;
   let errored = false;
   const data = {
@@ -20,13 +20,32 @@ export const login = async (email, password) => {
     data,
   };
   console.log(options);
-  try{
+  try {
     const response = await axios.post(apiUrlAuthLogin(), data);
     code = response.status;
-    if(code === 200) token = response.data.key;
+    if (code === 200) token = response.data.key;
   } catch {
     errored = true;
   } finally {
-    return {token, code, errored}
+    return { token, code, errored };
   }
 };
+
+/**
+ * 
+ * @modifies
+ */
+export const clearAuthTokenItem = () => localStorage.removeItem('authToken');
+
+/**
+ * 
+ * @returns {string} auth token
+ */
+export const getAuthTokenItem = () => localStorage.getItem('authToken');
+
+/**
+ * 
+ * @param {string} authToken 
+ * @modifies 
+ */
+export const setAuthTokenItem = (authToken) => localStorage.setItem('authToken', authToken)
